--- src/main/java/org/c4marathon/assignment/domain/account/entity/SettlementStatus.java
@@ -1,6 +1,0 @@
-package org.c4marathon.assignment.domain.account.entity;
-
-public enum SettlementStatus {
-	REQUESTED,
-	COMPLETED
-}
