+++ src/main/java/org/c4marathon/assignment/domain/account/entity/account/AccountRole.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.domain.account.entity;
+package org.c4marathon.assignment.domain.account.entity.account;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
