+++ src/main/java/org/c4marathon/assignment/domain/account/entity/settlement/Settlement.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.domain.account.entity;
+package org.c4marathon.assignment.domain.account.entity.settlement;
 
 import static jakarta.persistence.EnumType.*;
