+++ src/main/java/org/c4marathon/assignment/domain/account/entity/settlement/SettlementStatus.java
@@ -1,0 +1,6 @@
+package org.c4marathon.assignment.domain.account.entity.settlement;
+
+public enum SettlementStatus {
+	REQUESTED,
+	COMPLETED
+}
